# services/catalog-service/internal/http_server/docs/swagger.yaml
basePath: /v1
definitions:
  http_errors.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  models.Work:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      tumb:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: API catalog-service
  title: Catalog Service API
  version: "1.0"
paths:
  /works:
    post:
      consumes:
      - multipart/form-data
      description: Create a new work with title, description and image
      parameters:
      - description: Work title
        in: formData
        name: title
        required: true
        type: string
      - description: Work description
        in: formData
        name: description
        type: string
      - description: Work thumb
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Work'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_errors.ErrorResponse'
      summary: Creates a new work
      tags:
      - Works
  /works/{id}:
    get:
      description: Get a single work by its ID
      parameters:
      - description: Work ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Work'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http_errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_errors.ErrorResponse'
      summary: Get a work by ID
      tags:
      - Works
swagger: "2.0"